python_sonic_surface/requirements.txt
PySerial==3.5